refactor code, update imports
