+++ src/components/RandomSentence/RandomSentence.tsx
@@ -1,15 +1,15 @@
-import { useTransitionContext } from "../../utils/hooks/transitionContext";
-import { useWeatherContext } from "../../utils/hooks/weatherContext";
+import { useTransitionContext } from "@/context/transitionContext";
+import { useWeatherContext } from "@/context/weatherContext";
 
 const randomSentence = () => {
-    const transitionContextObject = useTransitionContext();
-    const weatherContextObject = useWeatherContext();
+  const transitionContextObject = useTransitionContext();
+  const weatherContextObject = useWeatherContext();
 
-    return ( 
-        <div className={`text-lg font-normal text-center mt-auto text-white ${transitionContextObject?.transition}-animation`}>
-            {weatherContextObject?.weatherObject?.currentWeather.randomSentence}
-            </div>
-     );
+  return (
+    <div className={`text-lg font-normal text-center mt-auto text-white ${transitionContextObject?.transition}-animation`}>
+      {weatherContextObject?.weatherObject?.currentWeather.randomSentence}
+    </div>
+  );
 }
- 
+
 export default randomSentence;
