+++ src/components/Search and Icon/IconWeather.tsx
@@ -1,26 +1,26 @@
 /// <reference types="vite-plugin-svgr/client" />
-import { useWeatherContext } from "../../utils/hooks/weatherContext";
-import { useTransitionContext } from "../../utils/hooks/transitionContext";
-import { useFirstRenderContext } from "../../utils/hooks/firstRenderContext";
-import getIconArray from "../../utils/getIconArray";
+import { useWeatherContext } from "@/context/weatherContext";
+import { useTransitionContext } from "@/context/transitionContext";
+import { useFirstRenderContext } from "@/context/firstRenderContext";
+import getIconArray from "@/utils/getIconArray";
 
 
 
-// Set up the Universal Icon Array
+// Set up the Icon Array
 const style = "w-full h-20 fill-white mt-[-8px] ml-[10px]"
 const iconArray = getIconArray(style);
 
 const IconWeather = () => {
-    const weatherContextObject = useWeatherContext();
-    const transitionContextObject=  useTransitionContext()
-    const firstRenderContextObject = useFirstRenderContext();
+  const weatherContextObject = useWeatherContext();
+  const transitionContextObject = useTransitionContext()
+  const firstRenderContextObject = useFirstRenderContext();
 
-    const Icon = iconArray[weatherContextObject?.weatherObject?.currentWeather.iconNumber ?? 10 ]
-    return ( 
-        <div className={`${firstRenderContextObject?.firstRender ? 'w-0' : 'w-36'} h-min ${transitionContextObject?.transition}-animation`}>
-            {Icon}
-        </div>
-     );
+  const Icon = iconArray[weatherContextObject?.weatherObject?.currentWeather.iconNumber ?? 10]
+  return (
+    <div className={`${firstRenderContextObject?.firstRender ? 'w-0' : 'w-36'} h-min ${transitionContextObject?.transition}-animation`}>
+      {Icon}
+    </div>
+  );
 }
- 
+
 export default IconWeather;
